[contactsd] Allow disabling transient SIM contact import.
This commit now changes the behaviour of the sim plugin depending
on the value of a configuration value.  If the value is "true"
then the sim plugin will transiently import sim contacts into
the phone addressbook automatically, otherwise it will not.
The default value assumed for the setting if it does not exist
is "true".